Require input trace
