+++ bin/predicate/Cargo.toml
@@ -16,3 +16,4 @@ logical-interface = { path = "../../logical-interface" }
 serde = { version = "1.0.203", features = ["derive"] }
 p3-util = { workspace = true }
 bin-common = { path = "../common" }
+hex = "0.4.3"
