+++ chips/Cargo.toml
@@ -28,6 +28,7 @@ tracing = "0.1.40"
 itertools = "0.13.0"
 getset = "0.1.2"
 serde = { version = "1.0", default-features = false, features = ["derive"] }
+derive_more = "0.99.18"
 
 [dev-dependencies]
 p3-uni-stark = { workspace = true }
